+++ src/command_modules/azure-cli-resource/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+unreleased
++++++++++++++++++++
+* `group deployment create`: Fixes issue where templates which lacked "parameters" or "resources" failed to deploy.
+
 2.0.11 (2017-07-27)
 +++++++++++++++++++
 * minor fixes